segmentation-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
torch==2.5.1
torchvision==0.20.1
pillow==11.0.0
numpy==2.1.3
